requirements.txt
# Файл: requirements.txt

fastapi
uvicorn
celery[redis]
sqlalchemy
psycopg2-binary
tinkoff-investments
torch
torchvision
pandas
scikit-learn
joblib
python-dotenv
newsapi-python
nltk
transformers
sentencepiece
python-telegram-bot
pandas-ta
alembic
psycopg2-binary
matplotlib
scipy
prophet
statsmodels
requests       
xmltodict      
optuna
xgboost
backtrader
shap
prometheus-fastapi-instrumentator
pytest
httpx
pytest-asyncio
pytest-mock
pytest-cov
pydantic-settings
